+++ CHANGELOG.md
@@ -28,6 +28,14 @@
 - [1.7.2 — 2025-10-31](#17172---2025-10-31)
 - [1.7.1 — 2025-10-29](#17171---2025-10-29)
 
+## 1.7.18 — 2025-12-18
+
+**Author:** Gabriel Glatz
+
+### Added
+
+- EventAgenda: adicionado novo componente.
+---
 ## 1.7.17 — 2025-12-17
 
 **Author:** Gabriel Glatz